requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via
    #   django
    #   django-cors-headers
cffi==1.16.0
    # via cryptography
cryptography==42.0.5
    # via djangorestframework-simplejwt
django==5.0.2
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
django-behaviors==0.5.1
    # via -r requirements.in
django-cors-headers==4.3.1
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-camel-case==1.4.2
    # via -r requirements.in
djangorestframework-simplejwt[crypto]==5.3.1
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt==2.8.0
    # via djangorestframework-simplejwt
pytz==2024.1
    # via djangorestframework
pyyaml==6.0.1
    # via -r requirements.in
sqlparse==0.4.4
    # via django
uritemplate==4.1.1
    # via -r requirements.in
whitenoise==6.6.0
    # via -r requirements.in
